\name{module_library}

\alias{module_library}

\alias{BioCroBML}

\title{BioCroBML}

\description{
  This R package represents a collection of BioCro modules. It is not intended
  to be used on its own; instead, the BioCro framework R package contains the
  essential functions required to interact with the modules in this library. For
  more information, see \code{\link[BioCro]{modules}},
  \code{\link[BioCro]{run_biocro}}, and the BioCro framework R package
  \href{https://github.com/biocro/biocro}{repository}.
}

\details{
  YOUR DESCRIPTION OF THE MODULE LIBRARY HERE.
}
